requirements.txt
flask
flask-cores
tensorflow
pandas
numpy
pillow
opencv-python

